# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/sign-in": {
            "post": {
                "description": "Auth User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Auth User",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SignIn"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "Registry User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Registry User",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SignUp"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "domain.SignIn": {
            "type": "object",
            "required": [
                "place"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "fio": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                }
            }
        },
        "domain.SignUp": {
            "type": "object",
            "required": [
                "email",
                "fathername",
                "firstname",
                "lastname",
                "login",
                "password",
                "salon"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "fathername": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "salon": {
                    "type": "integer"
                }
            }
        }
    }
}